Lib: Overlapper, OverlapArea, Platforms
First, Overlapper is a generic class that listens for supplied overlap
callback and runs supplied callbacks for enterings and leavings.
Overlaparea is a simple MapObjectable implementation.

Platforms now run on state machines and can be toggled as auto (moving
all the time) or not (moves after entering, with delay).

Calculating the distance is enough!  When the game freezes it causes it
to get off track, I still think some kind of interpolated snap should be
used, but another day.