Assets: big messy bundle... Will be saved by jdb
Among some additions here are the desert meta-maps, and the pieces that
go with it.  We call 'stamp' files those that are used as input in wang
or slice generation.  Also, fuck it, lets just version control xcf
files.  Doubt there will be much use in diffing them, but it stores its
own history!  And we will keep with the convention of keeping tiled
scripts in our asset commits.  We should expect these to be cleaned up
quite a bit with jdb.