Almost got negative offsets for addChunk
The problem is that our answer is to just reverse what is the base chunk
and what is the overlay in these instances, but then that makes bad
results when there is a negative offset that isn't larger than the
relevant dimension.

What I have to do is add something that crops the base I think, relative
to this difference.