First attempt at negative offsets for addChunk
This one fails where the negative offset is less than dimension of the
chunk, that is the base ends up overlaying the ol, which is not whats
intended in that situation.

Whats the new way (next commit)?

Create an assuredly big enough empty grid, and inject our two chunks
into it, where the negative offset ol chunk is the origin. It already
knows its negative, but not which one, so how can we get, say, an
assuredly big enough width? It's either, the base width + abs(negative
offset), the positive offset + ol width... and maybe thats
it? Why not the base width +  the positive offset? Because the ol has
/some/ width, so the third option will always cover this. Say there is
positive x offset thats larger than the base width, well its still
always just placing at least a block there from the ol, to make it one more.