feat: Cliff tilemap generation
This is what we will go with, extracted the procedure into a function
that just takes grids, and it produces grids.  Will rely on the calling
procedure to extract the grids, the replacement sets, and the applying
back to an actual tilemap.  It produces layers for each possible
altitude, and performs replacements per altitude.

Still need to add stairs/ramps.

N.B. next step is to make the wang terrain process follow the same interface.